requirements.txt
pdfplumber
Pillow
pytesseract
reportlab
gspread
google-auth
python-dotenv
numpy
opencv-python
requests
